cawachone/requirements.txt
asgiref==3.3.1
dateutils==0.6.12
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
mysqlclient==2.0.3
mysql-connector-python==8.0.23
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2021.1
paho-mqtt==1.5.1
PyMySQL==1.0.2
SQLAlchemy==1.3.23
six==1.15.0
sqlparse==0.4.1
uWSGI==2.0.19.1